change group to role

+++ interface/sa/sa_branch/sa_branch.go
@@ -8,19 +8,19 @@ import (
 
 // Repository :
 type Repository interface {
-	GetBySaBranch(ctx context.Context, branchID int16) (result sa_models.SaBranch, err error)
+	GetBySaBranch(ctx context.Context, branchID int) (result sa_models.SaBranch, err error)
 	GetList(ctx context.Context, queryparam models.ParamList) (result []*sa_models.SaBranch, err error)
 	CreateSaBranch(ctx context.Context, branchData *sa_models.SaBranch) (err error)
 	UpdateSaBranch(ctx context.Context, branchData *sa_models.SaBranch) (err error)
-	DeleteSaBranch(ctx context.Context, branchID int16) (err error)
+	DeleteSaBranch(ctx context.Context, branchID int) (err error)
 	CountBranchList(ctx context.Context, queryparam models.ParamList) (result int, err error)
 }
 
 // Branchcase :
 type Usecase interface {
-	GetBySaBranch(ctx context.Context, branchID int16) (result sa_models.SaBranch, err error)
+	GetBySaBranch(ctx context.Context, branchID int) (result sa_models.SaBranch, err error)
 	GetList(ctx context.Context, queryparam models.ParamList) (result models.ResponseModelList, err error)
 	CreateSaBranch(ctx context.Context, branchData *sa_models.SaBranch) (err error)
 	UpdateSaBranch(ctx context.Context, branchData *sa_models.SaBranch) (err error)
-	DeleteSaBranch(ctx context.Context, branchID int16) (err error)
+	DeleteSaBranch(ctx context.Context, branchID int) (err error)
 }
